+++ Dyna/docs/release-notes.md
@@ -7,6 +7,7 @@
 * Changed the Bang card.
 * Changed the Light It Up card.
 * Changed the Lock and Load (and by extension Custom Charge) cards.
+* Changed the Swift Charge.
 * Fixed Blastwave triggering at the wrong spot for attacks which move the enemy around.
 * Fixed the Bazooka Mod (Dyna-Isaac duo) artifact also triggering on enemy drones.